Enable dashboards.json property
This allows us to place dashboard configurations, which are stored
as JSON, onto the file system. Grafana will then pick these up
automatically, without the need for restarting. The persistent disk
(`/var/vcap/store`) was chosen as the location for these files as
it is the convention used by related tools, such as Graphite.