Create and update dashboards
This defines a property `grafana.dashboards`, which is a list of
objects containing the dashboard name and a content field. The
content field is the dashboard JSON minified so it fits on a single
line. Minified JSON was chosen to avoid converting to YAML, as
we would just have to convert it back to JSON again.

An extra hook has been added to the post-start script to take each
piece of JSON and create a JSON file. By placing this file in the
directory the Grafana UI is automatically updated with the new
dashboard.